Broke stateHandlers file into smaller files. Fix tests and typings for each file.
