+++ react/components/admin/pages/Form/stateHandlers/__fixtures__/newPage.ts
@@ -1,0 +1,21 @@
+const newPage = {
+  condition: {
+    allMatches: true,
+    id: '',
+    statements: [
+      {
+        error: '',
+        object: {
+          date: new Date('2019-02-01'),
+        },
+        subject: 'date',
+        verb: 'is',
+      },
+    ],
+  },
+  operator: 'all',
+  pageId: '',
+  template: '',
+}
+
+export default newPage
