+++ react/components/admin/pages/Form/stateHandlers/getLoginToggleState.ts
@@ -1,0 +1,9 @@
+import { State } from '../index'
+
+export const getLoginToggleState = (prevState: State) => ({
+  ...prevState,
+  data: {
+    ...prevState.data,
+    auth: !!prevState.data && !prevState.data.auth,
+  },
+})
